Create ScottPlot.AxisPanels.Experimental namespace
